129 - move data ingestion to pv_to_puml
