+++ proof_of_concepts/dash_app/app.py
@@ -14,7 +14,7 @@ from tel2puml.pv_to_puml.pv_to_puml import (
     pv_jobs_from_folder_to_puml_file,
     pv_jobs_from_folder_to_event_sequence_streams,
 )
-from tel2puml.data_pipelines.data_ingestion import (
+from tel2puml.pv_to_puml.data_ingestion import (
     get_graph_solutions_from_clustered_events,
 )
