+++ tests/tel2puml/test_logic_detection.py
@@ -6,7 +6,7 @@ from numpy.typing import NDArray
 
 
 from tel2puml.pv_event_simulator import generate_test_data
-from tel2puml.data_pipelines.data_ingestion import (
+from tel2puml.pv_to_puml.data_ingestion import (
     update_all_connections_from_data
 )
 from tel2puml.events import EventSet
