+++ tests/tel2puml/walk_puml_graph/test_walk_puml_logic_graph.py
@@ -16,7 +16,7 @@ from tel2puml.puml_graph import (
     PUMLGraph,
     PUMLOperatorNode,
 )
-from tel2puml.data_pipelines.data_ingestion import (
+from tel2puml.pv_to_puml.data_ingestion import (
     update_all_connections_from_clustered_events,
 )
 from tel2puml.events import events_to_markov_graph
